refactor(perl-dap): split inline values helpers (#0000)
